+++ database/migrations/2026_07_29_223046_create_pull_request_reviews_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('pull_request_reviews', function (Blueprint $table) {
             $table->id();
             $table->timestamps();
-            $table->integer('github_id')->unique();
+            $table->bigInteger('github_id')->unique();
             $table->foreignId('reviewer_id')->nullable()->constrained('users');
             $table->foreignId('pull_request_id')->nullable()->constrained('pull_requests');
             $table->enum('state', array_column(PullRequestReviewStateEnum::cases(), 'value'));
